fit: DirectorStorage объединён с FilmStorage
