+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.storage;
 
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.RatingMPA;
@@ -28,6 +29,18 @@ public interface FilmStorage {
     List<RatingMPA> findRatingMPAs();
 
     Optional<RatingMPA> findRatingMPAById(long ratingMPAId);
+
+    List<Director> findDirectors();
+
+    Optional<Director> findDirectorById(long directorId);
+
+    boolean isFindDirectorById(long directorId);
+
+    Optional<Director> createDirector(Director director);
+
+    Optional<Director> updateDirector(Director director);
+
+    boolean deleteDirectorById(Long directorId);
 
     List<Film> findSortFilmsByDirector(long directorId, String sortBy);
