+++ src/main/java/ru/yandex/practicum/filmorate/storage/memoryImpl/InMemoryFilmStorage.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.memoryImpl;
 
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.RatingMPA;
@@ -72,6 +73,30 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     public Optional<RatingMPA> findRatingMPAById(long ratingMPAId) {
         return Optional.empty();
+    }
+
+    public List<Director> findDirectors() {
+        return Collections.EMPTY_LIST;
+    }
+
+    public Optional<Director> findDirectorById(long directorId) {
+        return Optional.empty();
+    }
+
+    public boolean isFindDirectorById(long directorId) {
+        return false;
+    }
+
+    public Optional<Director> createDirector(Director director) {
+        return Optional.empty();
+    }
+
+    public Optional<Director> updateDirector(Director director) {
+        return Optional.empty();
+    }
+
+    public boolean deleteDirectorById(Long directorId) {
+        return false;
     }
 
     public List<Film> findSortFilmsByDirector(long directorId, String sortBy) {
